Refactor and restructured code
